requirements.txt
Flask==2.3.*
tensorflow==2.12.*
Pillow==9.5.*
